refactor: TextField common component
Add support for 'ref' and 'onBlur' props to support the focus on error feature for react-hook-form.
Completely rewrite the component type definition to force the use of the correct props types based on the value of the 'as' prop.
Add 'clsx' to calculate class names.
Simplify type calculations and behavior logic for "action" buttons.